# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation",
        "description": "API documentation for Laravel + React project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/teststable": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get all tests",
                "operationId": "c100efb2b7415b897a5dd3831e4b3ef7",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/getFilteredTests": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get filtered tests",
                "operationId": "50f2aa32ea99cdc8957fec7f6b2467b1",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter criteria",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered tests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/getcategorytests": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get tests by category",
                "operationId": "815e8a34f1e4630c0597ef3db4cb3f16",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tests filtered by category"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/delete/{id}": {
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Delete a test",
                "operationId": "ac24d18a72040b67e793547d51ff53eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the test to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/edittest/{id}": {
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Edit a test",
                "operationId": "21f0511e4b9e1d7a0c265e76bae22f07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the test to edit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_date",
                                    "end_date",
                                    "status",
                                    "level"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Active",
                                            "Inactive"
                                        ]
                                    },
                                    "level": {
                                        "type": "string",
                                        "enum": [
                                            "Beginner",
                                            "Intermediate",
                                            "Advanced"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/addtest": {
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Add a new test",
                "description": "* @OA\\Post(\n    path=\"/api/addtest\",\n    summary=\"Add a new test\",\n    tags={\"Tests\"},\n    security={{\"sanctum\":{}}},",
                "operationId": "e7583e07e24cf6a58be1e2081a398e82",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_date",
                                    "end_date",
                                    "status",
                                    "question",
                                    "level",
                                    "candidate",
                                    "product",
                                    "category",
                                    "template",
                                    "version"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test 1"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-10"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Active",
                                            "Inactive"
                                        ],
                                        "example": "Active"
                                    },
                                    "question": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "level": {
                                        "type": "string",
                                        "enum": [
                                            "Beginner",
                                            "Intermediate",
                                            "Advanced"
                                        ],
                                        "example": "Beginner"
                                    },
                                    "candidate": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "product": {
                                        "type": "string",
                                        "example": "Product A"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Category 1"
                                    },
                                    "template": {
                                        "type": "string",
                                        "example": "Template 1"
                                    },
                                    "version": {
                                        "type": "string",
                                        "example": "v1.0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Test added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Test added successfully."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Math Assessment",
                                                "start_date": "2025-04-01",
                                                "end_date": "2025-04-15",
                                                "status": "Active",
                                                "level": "Intermediate",
                                                "question": "101",
                                                "candidate": "John Doe",
                                                "product": "CBT Software",
                                                "category": "Mathematics",
                                                "template": "Default Template",
                                                "version": "v1.0"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error occurred while adding test",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error occurred while adding test: {error_message}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tests/count": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get total test count",
                "operationId": "11b15adb4d9aadd7c231e3aef2a43a6b",
                "responses": {
                    "200": {
                        "description": "Total number of tests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tests/active/count": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get active test count",
                "operationId": "011278adedd5cab63b009e913403413d",
                "responses": {
                    "200": {
                        "description": "Number of active tests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tests/export": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Export test data",
                "operationId": "c59cf62c0290406063b98f678556d2d7",
                "responses": {
                    "200": {
                        "description": "Exported test data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/searchTests": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Search tests",
                "operationId": "d0ef36348b8bc30e51b50444d495c9b7",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching tests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}